+++ src/main/java/use_case/registration/RegistrationDataAccessInterface.java
@@ -1,0 +1,8 @@
+package use_case.registration;
+
+import entity.User;
+import utility.exceptions.DuplicateUsernameException;
+
+public interface RegistrationDataAccessInterface {
+    void saveUser(User user) throws DuplicateUsernameException;
+}