+++ src/main/java/utility/exceptions/PasswordsDoNotMatchException.java
@@ -1,0 +1,7 @@
+package utility.exceptions;
+
+public class PasswordsDoNotMatchException extends Exception {
+    public PasswordsDoNotMatchException(String message) {
+        super(message);
+    }
+}