+++ src/main/java/utility/exceptions/WeakPasswordException.java
@@ -1,0 +1,7 @@
+package utility.exceptions;
+
+public class WeakPasswordException extends Exception {
+    public WeakPasswordException(String message) {
+        super(message);
+    }
+}